pyproject.toml
[project]
name = "content-creator"
version = "0.1.0"
description = "Google ADK를 활용한 콘텐츠 제작 에이전트"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "openai>=1.0.0",
    "litellm>=1.0.0",
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "pdfplumber>=0.10.0",
    "pypdf2>=3.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# content_creator 패키지 포함
packages = ["content_creator"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.black]
line-length = 100
target-version = ["py310"]

